remove interface/impl for WebACAuthorization in favor of a POJO
